+++ game/src/rendering/entities.rs
@@ -24,18 +24,18 @@ fn render_entity(
     if let Some(texture) = get_rendering_config().get_texture(sprite_key) {
         let source = item.texture_rect;
         let source_rect = Rectangle {
-            x: source.x as f32 * TILE_SIZE, 
-            y: source.y as f32 * TILE_SIZE,
-            width: source.w as f32 * TILE_SIZE,
-            height: source.h as f32 * TILE_SIZE,
+            x: source.x * TILE_SIZE, 
+            y: source.y * TILE_SIZE,
+            width: source.w * TILE_SIZE,
+            height: source.h * TILE_SIZE,
         };
 
         let frame = item.frame;
         let dest_rect = Rectangle {
             x: (frame.x - camera_viewport.x) * tile_scale,
             y: (frame.y - camera_viewport.y) * tile_scale,
-            width: frame.w as f32 * tile_scale,
-            height: frame.h as f32 * tile_scale,
+            width: frame.w * tile_scale,
+            height: frame.h * tile_scale,
         };
 
         d.draw_texture_pro(
@@ -51,8 +51,8 @@ fn render_entity(
         let dest_rect = Rectangle {
             x: (frame.x - camera_viewport.x) * tile_scale,
             y: (frame.y - camera_viewport.y) * tile_scale,
-            width: frame.w as f32 * tile_scale,
-            height: frame.h as f32 * tile_scale,
+            width: frame.w * tile_scale,
+            height: frame.h * tile_scale,
         };
 
         d.draw_rectangle_rec(dest_rect, Color::RED.alpha(0.5));
